requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
einops>=0.7.0
accelerate>=0.20.0
requests>=2.31.0
yt-dlp>=2023.7.6
fiftyone>=0.23.0
gradio>=4.0.0
gtts>=2.3.0  # Google Text-to-Speech for audio summaries

